fix: Change index type from u64 to usize in Args struct

+++ src/cli.rs
@@ -12,7 +12,7 @@ pub struct Args {
 
     /// Default position of the cursor 
     #[arg(short, long, default_value_t = 0)]
-    pub index: u64,
+    pub index: usize,
 
     #[arg(short, long, default_value_t = 10)]
     lines: u32
